configure score-p support
The previous change was no good, it was using the truncated F90 and CC
vars to generate the Makefiles.

I have removed much of the modifications to the Makefiles and now just
strip scorep from the problematic test of configure (fortran name
mangling).  I also replaced `scorep icc` with `icc` to fix another test.
(F-C linking)